Re-namespace Slack::BlockBuilder
